volume settings modified on notemodell. gui improvments.

+++ core/src/main/java/hu/boga/musaic/core/sequence/boundary/dtos/NoteDto.java
@@ -5,5 +5,5 @@ public class NoteDto {
     public int midiCode;
     public long tick;
     public long length;
-    public int velocity;
+    public double velocity;
 }
